requirements.txt
aioredis==1.3.1
amqp==5.0.5
appdirs==1.4.3
apturl==0.5.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
auditlog3==1.0.1
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
boto3==1.17.20
botocore==1.20.20
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.6
cupshelpers==1.0
daphne==3.0.1
dbus-python==1.2.16
deepmerge==0.1.1
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-3-jet==1.0.8
django-background-tasks==1.2.5
django-braces==1.14.0
django-celery-results==2.0.1
django-choices==1.7.1
django-ckeditor==6.3.0
django-compat==1.0.15
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2
django-debug-toolbar-request-history==0.1.3
django-dotenv==1.4.2
django-extra-fields==3.0.2
django-filter==2.4.0
django-js-asset==2.0.0
django-nose==1.4.7
django-picklefield==3.0.1
django-redis==4.12.1
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
factory-boy==3.2.0
Faker==6.5.0
fasteners==0.14.1
filelock==3.13.1
firebase-admin==4.5.2
flex==6.14.1
flower==0.9.7
future==0.18.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.3
grpcio==1.36.1
gunicorn==20.0.4
h11==0.12.0
hiredis==1.1.0
httplib2==0.19.0
humanize==3.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsondiff==1.2.0
jsonmerge==1.8.0
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
netifaces==0.10.4
newrelic==6.2.0.156
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.9
paramiko==2.6.0
pexpect==4.6.0
pika==1.2.0
Pillow==7.0.0
platformdirs==4.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.16
proto-plus==1.14.2
protobuf==3.15.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynamodb==5.0.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.34
requests==2.25.1
requests-aws4auth==1.0.1
requests-unixsocket==0.2.0
rfc3987==1.3.8
rollbar==0.15.2
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.41
sqlparse==0.4.1
ssh-import-id==5.10
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
systemd-python==234
text-unidecode==1.3
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==4.0.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.3
usb-creator==0.3.7
uvicorn==0.11.3
validate-email==1.3
vine==5.0.0
virtualenv==20.25.0
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
websockets==8.1
wsproto==0.15.0
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.interface==5.2.0
